google-adk-cookbook/requirements.txt
honeyhive[openinference-google-adk]
google-adk>=1.0.0
python-dotenv
openai
